backend/requirements.txt
anyio==4.12.1
asgiref==3.11.1
certifi==2026.1.4
Django==4.2.28
django-cors-headers==4.9.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
exceptiongroup==1.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
PyJWT==2.11.0
python-dotenv==1.2.1
sqlparse==0.5.5
typing_extensions==4.15.0
